Extract shared type refactoring helper for rename and move
