+++ plugins/org.eclipse.fordiac.ide.typemanagement/src/org/eclipse/fordiac/ide/typemanagement/refactoring/move/MoveTypeRefactoringParticipant.java
@@ -30,20 +30,15 @@ import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.OperationCanceledException;
-import org.eclipse.emf.ecore.EObject;
 import org.eclipse.fordiac.ide.model.IdentifierVerifier;
 import org.eclipse.fordiac.ide.model.helpers.PackageNameHelper;
-import org.eclipse.fordiac.ide.model.libraryElement.BlockFBNetworkElement;
-import org.eclipse.fordiac.ide.model.search.types.BlockTypeInstanceSearch;
-import org.eclipse.fordiac.ide.model.typelibrary.DataTypeEntry;
 import org.eclipse.fordiac.ide.model.typelibrary.TypeEntry;
 import org.eclipse.fordiac.ide.model.typelibrary.TypeLibraryManager;
 import org.eclipse.fordiac.ide.typemanagement.Messages;
 import org.eclipse.fordiac.ide.typemanagement.refactoring.ModelEdit;
 import org.eclipse.fordiac.ide.typemanagement.refactoring.ModelEditChange;
 import org.eclipse.fordiac.ide.typemanagement.refactoring.RefactoringUtil;
-import org.eclipse.fordiac.ide.typemanagement.refactoring.UpdateFBTypeModelEdit;
-import org.eclipse.fordiac.ide.typemanagement.refactoring.edit.DataTypeEditBuilder;
+import org.eclipse.fordiac.ide.typemanagement.refactoring.TypeRefactoringHelper;
 import org.eclipse.ltk.core.refactoring.Change;
 import org.eclipse.ltk.core.refactoring.CompositeChange;
 import org.eclipse.ltk.core.refactoring.RefactoringStatus;
@@ -103,14 +98,8 @@ public class MoveTypeRefactoringParticipant extends MoveParticipant {
 	@Override
 	public Change createChange(final IProgressMonitor pm) throws CoreException, OperationCanceledException {
 		final List<ModelEdit<?>> modelEdits = new ArrayList<>();
-		processTypeFiles(resource, destination.getFullPath(), (typeEntry, path) -> {
-			if (typeEntry instanceof final DataTypeEntry dtEntry) {
-				DataTypeEditBuilder.createStructuredDataTypeChanges(dtEntry, modelEdits,
-						DataTypeEditBuilder.getFullTypeName(path));
-			} else {
-				addInstanceChanges(modelEdits, typeEntry);
-			}
-		});
+		processTypeFiles(resource, destination.getFullPath(),
+				(typeEntry, path) -> TypeRefactoringHelper.addModelEditsForType(modelEdits, typeEntry, path));
 		return ModelEditChange.fromModelEdits(Messages.MoveTypeToPackage_UpdateInstances, modelEdits);
 	}
 
@@ -124,16 +113,6 @@ public class MoveTypeRefactoringParticipant extends MoveParticipant {
 		} else if (resource instanceof final IContainer container) {
 			for (final IResource member : container.members()) {
 				processTypeFiles(member, newPath.append(container.getName()), processor);
-			}
-		}
-	}
-
-	private static void addInstanceChanges(final List<ModelEdit<?>> modelEdits, final TypeEntry typeEntry) {
-		final List<? extends EObject> result = new BlockTypeInstanceSearch(typeEntry).performSearch();
-
-		for (final EObject eObject : result) {
-			if (eObject instanceof final BlockFBNetworkElement elem) {
-				modelEdits.add(new UpdateFBTypeModelEdit(elem, typeEntry));
 			}
 		}
 	}
